Creat costemission branch
This is the initial commit of the costemission branch of TEMOA by MAHTEP Group. It integrates in the model formulation the possibility to apply a cost to the emission of a specific emission commodity in a specific time period. It can be used to study scenarios including a carbon tax.